+++ vm/devices/tpm_resources/src/lib.rs
@@ -36,7 +36,7 @@ pub struct TpmDeviceHandle {
     /// BIOS GUID (for logging purposes)
     pub bios_guid: Guid,
     /// NVRAM size (default size if None)
-    pub nvram_size: Option<u64>,
+    pub nvram_size: Option<usize>,
 }
 
 impl ResourceId<ChipsetDeviceHandleKind> for TpmDeviceHandle {
